requirements.txt
Django>=4.0,<5.0
psycopg2-binary>=2.9
openai>=1.0
python-dotenv>=1.0
requests>=2.25
beautifulsoup4>=4.12
google-api-python-client>=2.100
django-cors-headers>=4.0
